// src/resources/boxes.ts
  export interface UnionMember0 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for a Linux box instance
     */
    config: UnionMember0.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Linux
     */
    type: 'linux';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember0 {
    /**
     * Configuration for a Linux box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Linux boxes
         */
        type: 'chromium' | 'firefox' | 'webkit';

        /**
         * Browser version string (e.g. '12')
         */
        version: string;
      }

      /**
       * Operating system configuration
       */
      export interface Os {
        /**
         * Supported operating system types for Linux boxes
         */
        type: 'linux' | 'windows' | 'macos';

        /**
         * OS version string (e.g. 'ubuntu-20.04')
         */
        version: string;
      }
    }
  }

  export interface UnionMember1 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for an Android box instance
     */
    config: UnionMember1.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Android
     */
    type: 'android';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }
// ...
export type BoxRetrieveResponse = BoxRetrieveResponse.UnionMember0 | BoxRetrieveResponse.UnionMember1;

export namespace BoxRetrieveResponse {
  export interface UnionMember0 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for a Linux box instance
     */
    config: UnionMember0.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Linux
     */
    type: 'linux';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember0 {
    /**
     * Configuration for a Linux box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Linux boxes
         */
        type: 'chromium' | 'firefox' | 'webkit';

        /**
         * Browser version string (e.g. '12')
         */
        version: string;
      }

      /**
       * Operating system configuration
       */
      export interface Os {
        /**
         * Supported operating system types for Linux boxes
         */
        type: 'linux' | 'windows' | 'macos';

        /**
         * OS version string (e.g. 'ubuntu-20.04')
         */
        version: string;
      }
    }
  }

  export interface UnionMember1 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for an Android box instance
     */
    config: UnionMember1.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Android
     */
    type: 'android';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember1 {
    /**
     * Configuration for an Android box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Android operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Android boxes
         */
        type: 'Chrome for Android' | 'UC Browser for Android';

        /**
         * Browser version string (e.g. '136')
         */
        version: string;
      }

      /**
       * Android operating system configuration
       */
      export interface Os {
        /**
         * Android operating system type
         */
        type: 'android';

        /**
         * Supported Android versions
         */
        version: '12' | '13';
      }
    }
  }
}
// ...
export type BoxDeleteResponse = BoxDeleteResponse.UnionMember0 | BoxDeleteResponse.UnionMember1;

export namespace BoxDeleteResponse {
  export interface UnionMember0 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for a Linux box instance
     */
    config: UnionMember0.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Linux
     */
    type: 'linux';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember0 {
    /**
     * Configuration for a Linux box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Linux boxes
         */
        type: 'chromium' | 'firefox' | 'webkit';

        /**
         * Browser version string (e.g. '12')
         */
        version: string;
      }

      /**
       * Operating system configuration
       */
      export interface Os {
        /**
         * Supported operating system types for Linux boxes
         */
        type: 'linux' | 'windows' | 'macos';

        /**
         * OS version string (e.g. 'ubuntu-20.04')
         */
        version: string;
      }
    }
  }

  export interface UnionMember1 {
    /**
     * Unique identifier for the box
     */
    id: string;

    /**
     * Configuration for an Android box instance
     */
    config: UnionMember1.Config;

    /**
     * Creation timestamp of the box
     */
    createdAt: string;

    /**
     * Expiration timestamp of the box
     */
    expiresAt: string;

    /**
     * The current status of a box instance
     */
    status: 'pending' | 'running' | 'stopped' | 'error';

    /**
     * Box type is Android
     */
    type: 'android';

    /**
     * Last update timestamp of the box
     */
    updatedAt: string;
  }

  export namespace UnionMember1 {
    /**
     * Configuration for an Android box instance
     */
    export interface Config {
      /**
       * Browser configuration
       */
      browser: Config.Browser;

      /**
       * CPU cores allocated to the box
       */
      cpu: number;

      /**
       * Environment variables for the box
       */
      envs: Record<string, string>;

      /**
       * Key-value pairs of labels for the box
       */
      labels: Record<string, string>;

      /**
       * Memory allocated to the box in MB
       */
      memory: number;

      /**
       * Android operating system configuration
       */
      os: Config.Os;

      /**
       * Storage allocated to the box in GB
       */
      storage: number;

      /**
       * Working directory path for the box
       */
      workingDir: string;
    }

    export namespace Config {
      /**
       * Browser configuration
       */
      export interface Browser {
        /**
         * Supported browser types for Android boxes
         */
        type: 'Chrome for Android' | 'UC Browser for Android';

        /**
         * Browser version string (e.g. '136')
         */
        version: string;
      }

      /**
       * Android operating system configuration
       */
      export interface Os {
        /**
         * Android operating system type
         */
        type: 'android';

        /**
         * Supported Android versions
         */
        version: '12' | '13';
      }
    }
  }
}